move all client configurations to separate class
